Update to C++17 and remove some boost usage
- Sweep out boost::algorithm::clamp for std::clamp
- Use <cstdint> instead of boost
- Use std::next / std::prev instead of boost
- Install a smaller subset of boost libraries for CI, and fix an error from building 32-bit with vcpkg

+++ source/widgets/playback/playbackwidget.h
@@ -63,7 +63,7 @@ private:
     Ui::PlaybackWidget *ui;
     QButtonGroup *myVoices;
 
-    const float MAX_ZOOM = 200, MIN_ZOOM = 25;
+    const double MAX_ZOOM = 200, MIN_ZOOM = 25;
 };
 
 #endif
